Made getMappingFilePath return empty string if an exception is thrown
To prevent an import from failing if a mapping file path can't be determined, any exception thrown in getMappingFilePath is logged and an empty string returned.

Background: Running a Java Application from within a jar file, the base directory can't be determined by calling ClassLoader.getSystemResource("") because it returns null.